requirements.txt
Django==2.2.20
matplotlib
numpy
pandas
paramiko
psutil
matplotlib
scipy
psycopg2-binary
djangorestframework
git+https://github.com/innogames/serveradmin.git